Mobile Game/requirements.txt
google-cloud-storage
apache_beam
python-dotenv
google-cloud-pubsub